Replacing spaces with "_" for samplenames. Also added nest ISA handling.
This means the sample name is modified for the entirety of the run; 
however, this does not impact the genelab raw read url as this is 
fetched directly from the file listing api.

Nest ISA means the ISA zip doesn't contain the i,s,a files directly but 
instead holds a folder that contains those folders.  The parser needs 
the folder that holds those files.